add dry_run option

+++ util/generate_bitstreams.py
@@ -38,6 +38,14 @@ def generate_bitstreams(args):
                 for collateral in os.scandir("/GarnetFlow/scripts/garnet/temp"):
                     os.rename(collateral.path, f"{entry.resolve()}/bin/{collateral.name}")
     else:
+        if args.dry_run:
+            for entry in args.apps:
+                name = entry.parts[-1]
+
+                if should_run_app(entry, args):
+                    logging.info(f"Generating bitstream for {entry}")
+            return
+
         with open(f"{args.app_root}/utilization.csv", "w") as f:
             w = csv.DictWriter(f, fieldnames=["name", "PE", "IO", "MEM", "REG"])
             w.writeheader()
